requirements.txt
numpy
pandas
scikit-learn
pygam
torch
dash
plotly
joblib
dash-bootstrap-components
flask-login
sqlalchemy
dash-auth
dash-extensions
bcrypt
reportlab
openpyxl
flask-babel
flask-mail
flask-wtf
pytest
pytest-cov
selenium
docker
gunicorn
flask-cors
flask-socketio
python-socketio
celery
redis
